requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
PyYAML==6.0.2

numpy==2.2.0
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2

mlflow==2.19.0

torch==2.5.1
transformers==4.47.0
datasets==3.1.0
